Refactor: last components
