+++ src/components/ProjectsList/index.tsx
@@ -1,8 +1,6 @@
-import { ArrowRight } from "../ProjectsNavigation/ArrowRight";
-import { ArrowLeft } from "../ProjectsNavigation/ArrowLeft";
 import classNames from "classnames";
 import { useEffect, useRef, useState } from "react";
-import { ProjectCard } from "../";
+import { ProjectCard, ArrowRight, ArrowLeft } from "../";
 
 type Projects = {
     id: string;
