+++ src/components/index.ts
@@ -1,20 +1,23 @@
-export * from './Header';
 export * from './BackButton';
-export * from './MobileMenuToggle';
-export * from './MobileMenu';
-export * from './LanguageSelector';
 export * from './BackgroundElement';
 export * from './Button';
-export * from './ProjectCard';
 export * from './Challenges';
-export * from './SectionTitle';
-export * from './SkillDescription';
 export * from './FigmaLink';
 export * from './Footer';
+export * from './Header';
+export * from './LanguageSelector';
+export * from './MobileMenu';
+export * from './MobileMenuToggle';
+export * from './ProjectCard';
 export * from './ProjectListFull';
 export * from './ProjectsList';
-export * from './ViewMore';
+export * from './ProjectsNavigation';
+export * from './SectionTitle';
 export * from './SkillBlock';
+export * from './SkillDescription';
 export * from './SkillLevel';
 export * from './SmallHeader';
 export * from './SubmitButton';
+export * from './TextArea';
+export * from './TextInput';
+export * from './ViewMore';
